ajout du backend du projet hotel

+++ controllers/clientsController.js
@@ -1,4 +1,7 @@
 const clientsService = require('../services/clientsService');
+const bcrypt = require('bcryptjs');
+const transporter = require('../config/nodemail');
+const jwt = require('jsonwebtoken');
 
 async function allClients(req, res) {
     try {
@@ -98,6 +101,9 @@ async function maxAmountSpent(req, res) {
 
 async function addClient(req, res) {
     try {
+        // Crypter le mot de passe
+        req.body.password = bcrypt.hashSync(req.body.password, 10);
+        
         const newClient = await clientsService.addClient(req.body);
         res.status(201);
         res.json(newClient);
@@ -109,6 +115,7 @@ async function addClient(req, res) {
 }
 async function updateClient(req, res) {
     try {
+        console.log(req.user);
         const updatedClient = await clientsService.updateClient(req.params.id, req.body);
         res.status(200);
         res.json(updatedClient);
@@ -121,6 +128,9 @@ async function updateClient(req, res) {
 
 async function deleteClient(req, res) {
     try {
+        if (req.user.role !== 'ADMIN') {
+            return res.status(403).json({ message: "Vous n'êtes pas autorisé à effectuer cette action" });
+        }
         const deletedClient = await clientsService.deleteClient(req.params.id);
         res.status(204);
         res.json(deletedClient);
@@ -131,16 +141,53 @@ async function deleteClient(req, res) {
     }
 }
 
+async function findMe(req, res) {
+    try {
+    const client = await clientsService.oneClient(req.user.id);
+    res.status(200);
+    res.json(client);
+    }
+    catch (error) {
+    console.error(error);
+    res.status(500);
+    res.json({ "message": "Une erreur est survenue lors de la récupération du client" });
+    }
+    }
 
 
+    async function passwordForget(req, res) {
+        try {
+            const client = await clientsService.findClientByEmail(req.body.email);
+            if (!client) {
+                return res.status(404).json({ message: 'Aucun utilisateur trouvé avec cet email' });
+            }
+            const token = jwt.sign({ id: client.id_client}, "SECRET", { expiresIn: '1h' });
+            await transporter.sendMail({
+                from: 'GENDRAULT Laura <lauragendrault@gmail.com',
+                to: client.email,
+                subject: 'Réinitialisation du mot de passe',
+                text: `Cliquez sur ce lien pour réinitialiser votre mot de passe: http://localhost:5173/reset-password/${token}`
+            });
+            res.status(200).json({ message: 'Un email vous a été envoyé pour réinitialiser votre mot de passe' });
+        } catch (error) {
+            console.error(error);
+            res.status(500);
+            res.json({ "message": "Une erreur est survenue lors de la récupération" });
+        }
+    }
 
-
-
-
-
-
-
-
+async function passwordReset(req, res) {
+    console.log(req.user);
+    try {
+        const client = await clientsService.updateClient(req.user.id, { password: bcrypt.hashSync(req.body.password, 10) });
+        res.status(200);
+        res.json(client);
+    } catch (error) {
+        console.error(error);
+        res.status(500);
+        res.json({ "message": "Une erreur est survenue lors de la récupération du client" });
+    }
+}
 
 
 
@@ -155,5 +202,8 @@ module.exports = {
     maxAmountSpent,
     addClient,
     updateClient,
-    deleteClient
+    deleteClient,
+    findMe,
+    passwordForget,
+    passwordReset
 };
